Add capabilities object to task payload
This commit also addresses an issue with hardcoding the number
of devices to find.  Instead, inspect the known location for them
and decide from that what devices could be used for tasks.

Also, devices should be specified in the task.payload.capabilities
that aligns with how we define privilege mode.