Bug 1171809 part 2: Implement purge cache events.
Let a build A creates a cached object directory Ca. Now let a build B
that reuses the cache Ca. If the build A fail and corrupts Ca, build B
will fail as well.

Enters the purge cache service. You can send a purge cache message to
clobber a build.

This commit implements the listening side of purge cache. docker-worker
listens for the purge cache event and remove the named cache. If the
cache is currently mounted, it schedules purging on garbage collection.