Bug 1220738 part 3: Update taskcluster-proxy temp credentials. r=garndt
When we reclaim a task, the credentials might get updated. When this
happens, we must send the new credentials to the taskcluster proxy, if
this feature is set to "true".

It uses the new taskcluster-proxy "Credentials" endpoint. The
credentials update integration test will fail until
https://github.com/taskcluster/taskcluster-proxy/pull/14 is merged.

We also had to update the unit test reclaim_test because now the Task
class extends EventEmitter.