Bug 1295184 part 2: Fix local live log test. r=garndt
Expiration time cannot be earlier than deadline. Also, docker sets local
live log expiration time to the min of the task expiration or bulkLog
expiration date (which is 60 minutes). As deadline is set to 60 minutes
in testworker.js, we need to set task expiration to 60 minutes to make
the test pass.