[deploy] Generate docker-worker.tgz out of tree. r=garndt
deploy/bin/build invokes tar to compress the repository, but it creates
a race condition because docker-worker.tgz, which is the name of the
compressed file, is stored inside the same directory as the repo, so tar
may try to include docker-worker.tgz in the compressed file, which is
docker-worker.tgz, then tar will detect that docker-worker.tgz was
modified while reading it.

The way to fix this is generating docker-worker.tgz out of the tree and
then moving it in, after tar is done.