[deploy] Generate AMI ids json file. r=garndt
Currently, the AMI ids for a newly generated image is only printed in
stdout. This means that, from a CI POV, the only way to retrieve it is
from output logs, and if a tool needs to save it, it needs to parse the
output and filter out the ids. Worse, the AMI info logs are multilined,
so we need a context free parser to safely extract AMI ids. This blocks
we from using the wonderful line oriented Unix text command line tools.

We solve this probem but storing the AMI ids in a json file that then
can be uploaded to a remote storage, like Taskcluster artifacts or S3
buckets.

packer has the option to generate a manifest json file with artifacts
information, and from this file we extract the AMI ids and store in a
new json file that's easier to read.