Bug 1307493: Apply exponential back off to sidecar startup. r=garndt
sidecar containers are the containers that start together with the task
container and that implement the task features. The service implemented
inside the sidecar cotainer may fail due to network hiccups, causing the
sidecar containers to fail to startup.

By applying startup retry with exponential back off we hope to fix these
random startup fails. Given the low probability of an startup fail, we
retry only 3 times.