[deploy] Use a permanent cot signing key. r=garndt
So far, each time we build a new AMI for docker-worker trusted
worker-types, we generate the cot signing key pair during AMI creation,
extract the public key and add it to the cot repo. The original mindset
is that the private key should never be in anyone's hands. A careful
threat modeling shows that the private key is exposed to anyone with AWS
credentials. Conclusion: the private is as safe as the AWS credentials
are. Therefore, generating a new key pair each time a AMI is built is
useless, and adds complexity to AMI rollout (push pub key to cot repo)
for no reward.

That said, move to use a permanent signing key instead of generating a
new one on each AMI build. We now have a third AMI set, called
"trusted", which ships the cot private key. The private key is stored in
passwordstore.

We also add a couple of utility scripts:
* gen-signing-key: generate a new signing key pair. Useful when we need
to rotate signing keys.
* cleanup-images: given a packer-artifacts.json file, it deregister the
referenced AMIs.