Update task ids for pull image test. r=garndt
The previous tasks used for this test expired, therefore we pick new
ones. As the task definitions are no longer available, we pick one of
the gecko in-tree generated zstd compressed images. For lz4, we add a
script that pulls the zstd image and create a new task with it
compressed as lz4.

As the lz4 image is somewhat bigger, we had to increase the test timeout
value.