Bug 1405683: Add option to disable worker types backup.
If, during the process of updating the worker types, an update fails,
the worker types set will be left in an in-between state where some are
updated and others not. If we run the script again, backups will be
overwritten with this undesirable state.

We add an option to update-worker-types script to not generate backup
file before updating.

We now also support the --test flag, which updates only ami-test and
ami-test-pv worker types.

These flags can also be passed to release.sh, which will forward them to
update-worker-types.