Configure the user of taskcluster-cli for temp cedentials
taskcluster-cli has the signin command that allows you to configure
temporary credentials given a set of scopes. The goal of this is to
avoid using permacreds with full access scopes in developers machines.

We change vagrant setup procedure to use this method. The
scopes.txt file contains the scopes required to execute tests,
and running:

eval $(taskcluster-cli signin --scope $(cat scopes.txt))

will yield an environment with proper Taskcluster credentials.