Support c5 instances
The upgraded AMIs correspond to the
ubuntu-trusty-14.04-amd64-server-20171115.1 releases. The HVM AMI
is ENA enabled, which means we can use c5 instances with it. (The
previous AMI was too old and didn't have ENA support.)

Because c5 instances have /dev/nvme* devices for block storage,
I had to update the /etc/default/docker-* provisioning scripts
to recognize those devices.

I did not update the AMIs listed in the app image because I
don't have the ability to generate "official" AMIs in the TC
account and with the proper embedded credentials. Someone will
need to build AMIs with this change and then update
deploy/packer/app.json with the result.