Upgrade to kernel 4.4 and use overlay2 storage driver
AUFS is the source of numerous performance problems and
filesystem consistency issues. Bug 1291940 has plenty of
details. The tl;dr is AUFS takes out a global kernel lock
for pretty much all I/O operations and concurrent I/O
grinds to a halt as you add more concurrency. Firefox builds
actually get slower as you use more cores. In addition, there
are consistency issues with AUFS that result in files randomly
not being evailable after write.

In this commit, I bump the kernel version from 3.13 to 4.4
(so overlayfs is usable), add support for the overlay2 storage
driver (added in Docker 1.12 - which we conveniently use),
and switch the default storage driver to overlay2.

The version of v4l2loopback was also bumped to the latest
in order to pull in fixes so it compiles with Linux 4.4.