Synchronize v4l2loopback installation
vagrant.sh now behaves like packages.sh with regards to installing
v4l2loopback. We now install a newer version from a Git repository.

While we're here, we also synchronize the gstreamer system packages that
are installed. And we do something a bit more sensible for defining
the kernel module settings.

Because the newly-built kernel modules can't be loaded until the next
reboot (because the kernel version is likely updated), we delete the
modprobe calls.