Remove dynamic task polling interval
The existing code assumes workers have a billing cycle. It dynamically
adjusts the task polling interval when workers approach a billing
cycle.

Since the code was written, AWS has switched to per second billing for
EC2 instances. So there is no such thing as a billing cycle anymore.
A dynamic polling interval that is based on billing cycles makes
no more sense. So this commit removes code and configuration for
that feature.