Don't take billing cycle into account when shutting down (bug 1424376)
EC2 instances used to have hourly billing. As of October 2017,
instances are now billed by the second.

Before, our termination strategy was based on hourly billing. We
kept instances alive for as much of a billing cycle as possible
in hopes they would do more work and we'd get our money's worth.
But with per second billing, this strategy actually results in
us wasting money because we're keeping instances around for longer
even if there is no work to do.

This commit removes all the code around billing cycles influencing
shutdown and replaces it with a simpler "shutdown after N seconds of
idle" feature.

Because uptime no longer influences shutdown, various references
to this variable have been removed throughout the code, especially
in tests.

When a worker comes online, its idle timeout is triggered. If no
tasks arrive before that timeout, the worker shuts down without
doing any work.

When a worker processes a task, its idle timer is reset. As soon
as the worker is idle, its shutdown timer is set and the worker
will shut down if no work arrives before that timer fires.

The default idle shutdown timer is 5 minutes. This was chosen
somewhat arbitrarily. We want machines to be waiting for work so
we don't need to wait for machines to spawn when new work is created.
But we don't want them waiting too long because that wastes money.
I figured 5 minutes is a good value because I'm assuming it is
roughly how long it takes for new workers to be provisioned. We
probabably prefer for workers to remain idle for longer than the
provisioning time. Otherwise we're wasting more time provisioning
workers. The value of an existing worker is worth more than a new
worker because it has paid the provisioning cost and it likely
has populated caches that can speed up scheduled tasks.