Consolidate provisioning code into packages.sh
vagrant.sh and packages.sh are near copies of each other. Keeping them
in sync is work.

This commit teaches vagrant.sh to source packages.sh and removes the
now redundant code from vagrant.sh.

In the few places where the code varied, we introduce a
VAGRANT_PROVISION variable to determine if we're running in Vagrant mode
and to skip actions that don't make sense there.