Rename docker-overlay2 to docker
We no longer have a need to paramaterize the init scripts because
of varying filesystems. Plus, using a separate init script for
each filesystem type led to divergence in the init scripts, making
them harder to change.

If we want to paramaterize init scripts in the future, we can write
the filesystem name to a variable in the script and have the script
take variable action depending on the filesystem being used. Or
something like that.