Remove diamond
Looking at cloud-init logs on a server, diamond was failing to
start.

Diamond is related to data capture for influxdb. According to
dustin, influxdb is dead.

This commit removes references to diamond. cloud-init should now
start fully working again. FWIW, the diamond failure was the last
line in cloud-init's per-boot init.sh. So I /think/ this failure
didn't impact other parts of cloud-init. But there is that
possibility. We should compare the cloud-init logs after this
change and make sure everything is happy.