Wait for cloud-init before starting docker
cloud-init does low-level things like configure networking
and storage. We should wait for cloud-init to finish before we
attempt to start docker.

cloud-init touches a boot-finished file when it has finished.
The file will be touched in case of error. That's OK... for now.
We simply wait on that file to come into existence before proceeding
with startup.

Ideally, we'd record this dependency in /etc/init/docker. However,
that file is provided by the docker package. We don't want to
overwrite it. I don't think Upstart lets us define override files.
We already have an /etc/default/docker file/script. So we abuse
it to inject hacky dependency management.