Provide a script to create github releases
The old release.sh script is now called deploy.sh. The new release.sh is
intended to create a Github release of docker-worker.

By running it, it creates a tag which name is based on current date
and time, and uploads the worker-types-backup.json and
docker-worker-amis.jsonjson files.

release.sh is intended to run right after deploy.sh. The reason to put
them in two scripts is because we may want to test the new rolled out
AMIs first before doing a release, or we are just rolling out AMIs on
testing worker-types during development.

Notice the release script requires an environment variable called
DOCKER_WORKER_GITHUB_TOKEN with a user personal token set.