Fix parallel task run test
Parallel task test infers if tasks ran in parallel by testing if they ran
in a bounded amount of time. This methodology is weak because the time a
task takes to complete may vary dramastically because the virtual
machine where the tests execute may be under heavy load.

Theoretically we should apply some statistics by repeating the tests a
certain number of times, removing the outliers and calculate the average
time along with its variance, and then check if the results are
acceptable.

As the development time urges, we take a simpler approach: we run the
tests until the time spent to execute the tasks is bounded, at a maximum
of 10 tries, and if at least one try succeeds, the test passes.

As the parallelism of a task, by itself, is something deterministic, if
we get one green iteration, it is enough.