Adjust uploadToS3 timeout to avoid saturation
Upload failures are, in most of the cases, due to a very short time
external problem; it doesn't make sense to wait for several minutes
to try to upload again. That said, we set the maximum back off time
to 30 seconds.

As we retry 10 times, we must adjust the exponential factor to make
the timeout at most 30 seconds in worst case to avoid timeout
saturation.