Bug 1527107: Fix CVE-2019-5736
docker-ce >= 18.09.2 fixes this vulnerability. Unfortunatelly, docker
channel doesn't distribute this version for Ubuntu trusty. As trusty
support is expiring too, we update the Ubuntu version in the AMI to
bionic.

The default kernel for bionic in EC2 is the linux-image-aws, but it
doesn't come with V4L2, necessary for some tests. We then switch to the
generic kernel.

Notice bionic uses systemd, and we have to upgrade startup scripts to
cope with it. We had to modify some systemd scripts to break dependency
cycles.