Delegate instance shutdown to the host configuration
Now each host configuration must implement a "shutdown" function. In the
case of worker-runner, it will call worker-manager removeWorker API and,
if it fails, falls back to a "system shutdown".

This must fix the problem in which instances start to shutdown but
stalls in the middle of the process, hanging forever.

Depends on: https://github.com/taskcluster/taskcluster/pull/2425